move container resources to AppAzurermStack
